requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
pydantic==2.5.2

# Development tools
black==23.12.1
flake8==7.0.0
pytest==7.4.3
pytest-cov==4.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2 